+++ tests/testthat/test-linalg.R
@@ -210,9 +210,15 @@ test_that("mat_pow", {
   skip_if_no_sympy()
   
   M <- matrix_(c("1", "a", "a", 1), 2, 2)
-  Msqrt <- mat_pow(M, 1/2)
   
-  expect_equal(as.character(Msqrt), 
-               "Matrix([[sqrt(1 - a)/2 + sqrt(a + 1)/2, -sqrt(1 - a)/2 + sqrt(a + 1)/2], [-sqrt(1 - a)/2 + sqrt(a + 1)/2, sqrt(1 - a)/2 + sqrt(a + 1)/2]])")
+  if ("pow" %in% names(M$pyobj)) {
+    Msqrt <- mat_pow(M, 1/2)
+    
+    expect_equal(as.character(Msqrt), 
+                 "Matrix([[sqrt(1 - a)/2 + sqrt(a + 1)/2, -sqrt(1 - a)/2 + sqrt(a + 1)/2], [-sqrt(1 - a)/2 + sqrt(a + 1)/2, sqrt(1 - a)/2 + sqrt(a + 1)/2]])")
+  } else {
+    expect_error(mat_pow(M, 1/2))
+  }
+  
 })